Cache Relevancy
This is experimental, but items should be cached based on how frequent they are accessed not a static final cache time of an hour.

This applies only to GET right now until further testing, it allows the user to get new data for specific posts. Posts that do not get engagement frequently will not update within the 15 to 45 min time window.