Adding Relevant Caching Mechanics
Before postly was caching each item with the same ttl of 1 hour, that is way too low for frequent and infrequent visited items.


Postly now increments a counter of number of visits to an object, that object then is stored and has an expiration time relevant to the number of times it has been accessed. Allowing posts to quickly update or save us time by not updating for a set time frame since it has not been accessed!